Introduce BKDIndexInput abstraction
