+++ lucene/core/src/java/org/apache/lucene/util/bkd/BKDIndexInput.java
@@ -1,0 +1,90 @@
+package org.apache.lucene.util.bkd;
+
+import java.io.IOException;
+import org.apache.lucene.index.PointValues.IntersectVisitor;
+import org.apache.lucene.util.BytesRef;
+
+/**
+ * Abstraction of a block KD-tree that contains multi-dimensional points in byte[] space.
+ *
+ * @lucene.internal
+ */
+public interface BKDIndexInput {
+
+  /** BKD tree parameters */
+  BKDConfig getConfig();
+
+  /** min packed value */
+  byte[] getMinPackedValue();
+
+  /** max packed value */
+  byte[] getMaxPackedValue();
+
+  /** Total number of points */
+  long getPointCount();
+
+  /** Total number of documents */
+  int getDocCount();
+
+  /** Create a new {@link IndexTree} to navigate the index */
+  IndexTree getIndexTree();
+
+  /** Create a new {@link LeafIterator} to read all leaf nodes */
+  LeafIterator getLeafTreeIterator() throws IOException;
+
+  /** Basic operations to read the BKD tree. */
+  interface IndexTree extends Cloneable {
+
+    /** Clone, but you are not allowed to pop up past the point where the clone happened. */
+    IndexTree clone();
+
+    /** Return current node id. */
+    int getNodeID();
+
+    /**
+     * Navigate to left child. Should not be call if the current node has already called this method
+     * or pushRight.
+     */
+    void pushLeft();
+
+    /**
+     * Navigate to right child. Should not be call if the current node has already called this
+     * method.
+     */
+    void pushRight();
+
+    /** Navigate to parent node. */
+    void pop();
+
+    /** Check if the current node is a leaf. */
+    boolean isLeafNode();
+
+    /** Check if the current node exists. */
+    boolean nodeExists();
+
+    /** Get split dimension for this node. Only valid after pushLeft or pushRight, not pop! */
+    int getSplitDim();
+
+    /** Get split dimension value for this node. Only valid after pushLeft or pushRight, not pop! */
+    BytesRef getSplitDimValue();
+
+    /** Get split value for this node. Only valid after pushLeft or pushRight, not pop! */
+    byte[] getSplitPackedValue();
+
+    /** Return the number of leaves below the current node. */
+    int getNumLeaves();
+
+    /** Visit the docs of the current node. Only valid if isLeafNode() is true. */
+    void visitDocIDs(IntersectVisitor visitor) throws IOException;
+
+    /** Visit the values of the current node. Only valid if isLeafNode() is true. */
+    void visitDocValues(IntersectVisitor visitor) throws IOException;
+  }
+
+  /** Navigate the leaf nodes of the tree. */
+  interface LeafIterator {
+
+    /** Visit the next leaf node. If the tree has no more leaves, it returns false. */
+    boolean visitNextLeaf(IntersectVisitor visitor) throws IOException;
+  }
+}
