+++ lucene/core/src/test/org/apache/lucene/util/bkd/Test2BBKDPoints.java
@@ -72,7 +72,7 @@ public class Test2BBKDPoints extends LuceneTestCase {
 
     IndexInput in = dir.openInput("1d.bkd", IOContext.DEFAULT);
     in.seek(indexFP);
-    BKDReader r = new BKDReader(in, in, in);
+    BKDReader r = new BKDReader(new BKDDefaultIndexInput(in, in, in));
     CheckIndex.VerifyPointsVisitor visitor = new CheckIndex.VerifyPointsVisitor("1d", numDocs, r);
     r.intersect(visitor);
     assertEquals(r.size(), visitor.getPointCountSeen());
@@ -121,7 +121,7 @@ public class Test2BBKDPoints extends LuceneTestCase {
 
     IndexInput in = dir.openInput("2d.bkd", IOContext.DEFAULT);
     in.seek(indexFP);
-    BKDReader r = new BKDReader(in, in, in);
+    BKDReader r = new BKDReader(new BKDDefaultIndexInput(in, in, in));
     CheckIndex.VerifyPointsVisitor visitor = new CheckIndex.VerifyPointsVisitor("2d", numDocs, r);
     r.intersect(visitor);
     assertEquals(r.size(), visitor.getPointCountSeen());
